Conflicts:
	config/routes.rb

+++ db/migrate/20130725054737_drop_join_expert_table.rb
@@ -1,6 +1,6 @@
 class DropJoinExpertTable < ActiveRecord::Migration
 	def up
-		drop_table :join_experts
+		drop_table :join_experts if self.table_exists?("join_experts")
 	end
 
 	def down
